[US_UT] Fix IC supervision level mapping (Recidiviz/recidiviz-data#40831)
## Description of the change

I misunderstood Hugo's request! We actually want the `COMPACT IN`
strings to be in the raw text but not overwrite the supervision level
enum when they are present. This keeps all of the changes from
https://github.com/Recidiviz/recidiviz-data/pull/40818, but takes out
the check for the `COMPACT IN` strings in the custom parser. That means
that people who are `COMPACT IN` will still have their actual
supervision levels reflected in the enum field, and their interstate
status reflected in the raw text field.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 46afa595552fa864ed137a4ba1b3dc2cbe2a822d