[US_AZ] Filter post-migration updates to old supervision periods (Recidiviz/recidiviz-data#40728)
## Description of the change

I was using the `UPDT_DTM` value on the `DPP_EPISODE` table to glean the
dates on which supervision level changes happened; however, this only
works for changes that happened _after_ the system migration on
2019-11-30. All rows for periods that ended before that migration were
both "created" and most recently updated on the date of the migration,
so using the `UPDT_DTM` value was holding periods open for an
unreasonably long time or adding an update well after the period was
closed.

The implication of this is that for supervision periods that started and
ended prior to the system migration on 2019-11-30, we will only be able
to observe the last-assigned supervision level a client had. We can see
all of the supervision levels that person ever had in the raw data, but
the `CREATE_DTM` and `UPDT_DTM` fields are all identically overwritten
to '2019-11-30', making the updates unable to be located in time.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a0cac8fa66271fd4270f9297aa733599f345cb42