[US_TN] Filter out certain non-physical movements when building IPs (Recidiviz/recidiviz-data#39278)
## Description of the change

Some IPs in TN are getting terminated/broken up due to PRDI-COURT or
CCDI-COURT movements, but these "movements" often are used to indicate
sentence closures rather than physical movements and shouldn't be
treated as movements when building IPs. However, there are cases where a
PRDI-COURT or CCDI-COURT movement _does_ indicate an actual release
event and _should_ be considered when building IPs. The movements we
want to filter out can be identified as follows:
- Movement type PRDI or CCDI, movement reason COURT
- "From" location doesn't match the "to" location of the previous
movement
- "To" location doesn't match the "from" location of the next movement
- "To" location of the previous movement _does_ match the "from"
location of the next movement
So if someone has a movement from A to B, then a PRDI-COURT movement
from B to C, then a movement from C to D, we wouldn't filter out the
PRDI-COURT movement because the locations match with the adjacent
movements. IF someone has a movement from A to B, then a PRDI-COURT
movement from X to Y, then a movement from B to C, then we would filter
out the PRDI-COURT movement, because its locations don't match with the
adjacent movements, and the adjacent movements do have a matching
location (if the PRDI-COURT movement is removed, then we're left with a
movement from A to B and then from B to C, which is a logical sequence).

This change will fix 81 errors in
incarceration_population_person_level_external_comparison, but will
introduce 17 new ones. Most of the new errors resulting from this change
are for people serving spans of temporary custody that we previously
weren't ingesting, and the external accuracy dataset doesn't include
them during these spans. This kind of error seems acceptable, since a)
there are very few, b) the external accuracy dataset seems to be
inconsistent at counting people who are serving temporary custody, and
c) based on the raw OffenderMovement data, it does look like these
people are in fact incarcerated at these times, even though they aren't
reported in the external accuracy dataset.

I've confirmed that this change will fix the issue where the people
flagged in Recidiviz/recidiviz-data#37998 have their incarceration periods incorrectly
interrupted by PRDI-COURT and CCDI-COURT movements.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#37998

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 94a6f58db87d17f79bbc66cac0851d6a22e8660f