[US_TN] Add IP normalization logic to improve continuity of custody periods (Recidiviz/recidiviz-data#40490)
## Description of the change

Adds normalization logic dealing with custodial_authority, using the
newly added T/P flags in admission_reason_raw_text and
release_reason_raw_text. This change adds 2 extra steps to
infer_additional_periods:

- Override the custodial authority for certain periods (if the custodial
authority changes in a period that has "-T" in the admission reason, and
the "T" period comes after a period with "-P" in the admission reason,
then the "T" period will be updated to use the custodial authority from
the "P" period).

- Infer periods to fill the gaps between certain temporary transfer
periods. Periods will only be inferred between 2 periods if the 2
periods have a gap (i.e. release date for the first period is before the
admission date for the second), if the 2 periods have the same custodial
authority, the 1st period has a "-T" in the release reason and the 2nd
has a "-T" in the admission reason, and the release reason for the 1st
period and the admission reason for the 2nd period both include paired
movement codes indicating a transfer to another jurisdiction or a work
release.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#39694

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f14a265ee2c60ffb95e48b93385d366af1279770