[Validations] Do not run raw data stable counts validation on code files (Recidiviz/recidiviz-data#40859)
## Description of the change

There are _lots_ of code files in Utah that have fewer than 10 rows.
Every time a single row gets added or removed, the raw data stable
historical counts validation blocks import. The list of exclusions in
was getting quite long for UT code files alone, so I am proposing that
we do not run this validation on code files. This PR adds a check for
the `is_code_file` flag to the `validation_applies_to_table` function to
accomplish that. Submitting for doppler review!

An alternative solution would be not to run this validation on files
that have fewer than 10 (or 20?) rows, but that may cause us to miss
more consequential changes.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c460e3630042ff78b8f0d6d2f3e6a68b29be227c