[Bugfix] Handle files with no BQ file_id in deprecate_ingest_bucket_raw_files (Recidiviz/recidiviz-data#40871)
## Description of the change

It's possible for a file to be stored in
`direct_ingest_raw_gcs_file_metadata` without having a set `file_id` in
the `direct_ingest_raw_big_query_file_metadata`. This PR fixes a bug
where we'd crash if we were only deprecating files that did not have
file_ids but did have gcs_file_ids.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See thread here:
https://recidiviz.slack.com/archives/C028X32LRH7/p1744728503790469

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: eac0adc86520ec338dd8b32eca8609ba521c94c0