[Validations] Bump thresholds for user metrics checks (Recidiviz/recidiviz-data#40170)
## Description of the change

Set validation overrides for a few failing validations that are not
concerning at their current rates:
- ME active users in officer monthly metrics vs impact reports: this is
a small numbers issue, 2 people don't line up and that's over the
current 5% threshold, so I bumped it to 10%
- MI and TN active usage without logins: this is a small edge case with
snooze events occurring late on the last day of the month with no logins
from the user in the following month. Right now it's only impacting 3
users from late 2024 so it shouldn't be noticeable in any external
reports. However, I kept the overrides tight to only cover the current
errors so we will get alerted if/when it shows up in a more recent
month.

## Related issues

Relates to Recidiviz/recidiviz-data#38769
Relates to Recidiviz/recidiviz-data#39969
Relates to Recidiviz/recidiviz-data#39457

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a920047e3435facf01808c1bcb7375d333eba09a