Refactor `is_past_completion_date_criteria_builder` to migrate criteria off v1 infra (Recidiviz/recidiviz-data#40092)
## Description of the change

This PR refactors the `is_past_completion_date_criteria_builder` to
reference `person_projected_date_sessions` (v2 infra) instead of
`incarceration_projected_completion_date_spans` /
`supervision_projected_completion_date_spans` (v1 infra). This update
migrates 25 criteria off of the sentencing v1 infra (19 criteria that
reference this function directly and 6 that reference it through another
criteria or task). The full list of impacted criteria is
[here](https://docs.google.com/spreadsheets/d/1Aai8iP0RvUEzue561cNKEPY-osJclctfdsK6HaeEinw/edit?usp=sharing).

Validation of the downstream impact is done in this
[notebook](https://colab.research.google.com/drive/1ljPnbS-B8lAILROVpzlyiJEzl9SnV3iw?usp=sharing)
and I look at changes in current meets_criteria counts at the criteria
level and changes in is_eligible counts at the task level. Note that the
counts that meet the a criteria impacted by this PR do change quite a
bit. However, this is because the projected completion date spans views
had logic that clips spans to incarceration and supervision session end
dates, which `person_projected_date_sessions` does not do. However, that
should have no impact on the task eligibility, since that exclusion (of
cases where a person has an active sentence but is not in the relevant
compartment), should be made by the candidate population. The following
is a screenshot from the notebook of tasks that have a >1% change in
current eligible population.

<img width="1079" alt="Screenshot 2025-04-02 at 2 29 17 PM"
src="https://github.com/user-attachments/assets/a2455b3a-a65a-405a-9371-6f94da946efe"
/>

The large drop for TN annual classification review is unrelated to this
PR. I checked with Damini, and big drops in eligible counts are expected
at the start of the month, and I believe this just has to do with the
timing of the sandbox run.

ND Transfer to CPP request has an 8% increase in eligible population. In
the few examples that I checked, this was due to sentences that are
imposed in the future, whose projected dates are no longer considered
relevant as part of this change. As discussed in OH, this opportunity is
not yet live and we are comfortable with this.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Justine Kunz <justine@recidiviz.org>
GitOrigin-RevId: 1db7e0bb04ade324d2dedcd99ef79f81cf9da5b0