[US_IX] Update StateStaff tests to prep for person staff periods (Recidiviz/recidiviz-data#40779)
## Description of the change

Creating tests for PersonStaffPeriods went well, but we needed related
staff to be ingested for the integration test.
This PR:
- updates the raw data config to allow automatic fixture generation
- updates the generation skip to allow empty files. In this case the
staff we were interested in a staff ID from the new data system, and so
write an empty file for the old data system
- leaves a comment adding some context on the existing integration test

This is really my first time dealing with person/staff relationships,
and definitely didn't have it properly in my mental model when doing
Janet design. The next PR that hydrates the new entity will have some
TODOs associated with the integration test, as hydrating the many person
staff relationships across all tests would be too cumbersome.

PII is in https://go/fixture-pii

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3ae1f89d7eb90a2deea7a4f6b837ea8597735403