[Insights] compare districts case-insensitively for outliers emails (Recidiviz/recidiviz-data#40908)
## Description of the change

In outliers emails, we cc the district manager of the supervisor's
district onto emails, so we need to join the fields in the two DBs. In
https://github.com/Recidiviz/recidiviz-data/issues/35349 we learned that
due to some details with how things are ingested, the locations might
have different cases for supervisors/DMs. This fixes the two places that
join on district to compare case-insensitively. Validation view loaded
to sandbox at `dana_location_validation`; confirmed that the PA row is
no longer there.

cc @brendali121

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35349

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0d5ffbed52a4308da6b3c59ff03e1c82d346210d