[US_AZ] Fix bug in nonstate sentence view (Recidiviz/recidiviz-data#40978)
## Description of the change

The ARS_ID values in raw data are sometimes in the form `62.0` instead
of `62` like I tested in my unit tests. This is making it so that
charges from the county do not connect to a description in the
`nonstate_sentence` ingest view, and are then filtered out of the
comprehensive list of charges in `state_charge_v2`.

This PR standardizes the `ARS_ID` formatting in the `JOIN` statement so
that it will always join, and these offenses will be included in the
final results. It also changes the formatting of the `ARS_ID` for some
rows in the unit test to ensure this works.

This PR also allows for charges to be ingested when the `DISPOSITION_ID`
is `NULL`. Time Comp has let us know they expect to see some of these
charges having an effect on eligibility, so we assume they led to a
conviction.

[successful sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-04-18_05_42_39-18427270035122658773;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

I also ensured that we can see every disqualifying charge that was
reported to be missing from the tool in Recidiviz/recidiviz-data#40941.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ea5037a4af587e6c385278d03b2a7a235109ef27