[Workflows] Update `snoozeCompanionOpportunityTypes` to be nullable in OpportunityConfiguration schema (Recidiviz/recidiviz-data#41237)
## Description of the change

Update `snoozeCompanionOpportunityTypes` to be nullable in
OpportunityConfiguration schema.

Throwing this temporary solution out there to unblock incase folks are
more comfortable with this and want more time to review the solution in
Recidiviz/recidiviz-data#40997.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: Mahmoud <mahmoud@Mahmouds-MacBook-Pro.local>
GitOrigin-RevId: e5b01cf358e351ef015801e8953f43f9584feb70